requirements.txt
python-barcode
pillow
flask